infra/compliance-scanners/lambdas/security_assistant/requirements.txt
# SecurityAssistant Lambda Dependencies

# LangChain core
langchain-core==0.3.29
langchain-community==0.3.13

# LLM — Groq (fast inference, llama-3.3-70b)
langchain-groq==0.2.3
langchain-openai
# Embeddings — Google Gemini
langchain-google-genai==2.0.8
google-generativeai==0.8.3

# Vector store — Pinecone
langchain-pinecone==0.2.0
pinecone-client==5.0.1

# Required by LangChain
pydantic==2.10.6
pydantic-core==2.27.2